[MOVE/#355] 프로필 수정 뷰 분리
