+++ app/src/main/AndroidManifest.xml
@@ -145,11 +145,11 @@
             android:exported="false"
             android:screenOrientation="portrait" />
         <activity
-            android:name=".presentation.main.profile.mod.SchoolProfileModActivity"
+            android:name=".presentation.main.profile.schoolmod.SchoolProfileModActivity"
             android:exported="false"
             android:screenOrientation="portrait" />
         <activity
-            android:name=".presentation.main.profile.mod.UnivProfileModActivity"
+            android:name=".presentation.main.profile.univmod.UnivProfileModActivity"
             android:exported="false"
             android:screenOrientation="portrait" />
         <activity
