+++ app/src/main/java/com/el/yello/presentation/main/profile/detail/ProfileDetailActivity.kt
@@ -1,6 +1,5 @@
 package com.el.yello.presentation.main.profile.detail
 
-import android.app.Activity.RESULT_OK
 import android.content.Intent
 import android.os.Bundle
 import androidx.activity.result.contract.ActivityResultContracts
@@ -11,8 +10,8 @@ import androidx.lifecycle.lifecycleScope
 import com.el.yello.R
 import com.el.yello.databinding.ActivityProfileDetailBinding
 import com.el.yello.presentation.main.profile.info.ProfileFragment.Companion.TYPE_UNIVERSITY
-import com.el.yello.presentation.main.profile.mod.SchoolProfileModActivity
-import com.el.yello.presentation.main.profile.mod.UnivProfileModActivity
+import com.el.yello.presentation.main.profile.schoolmod.SchoolProfileModActivity
+import com.el.yello.presentation.main.profile.univmod.UnivProfileModActivity
 import com.el.yello.util.Utils.setImageOrBasicThumbnail
 import com.el.yello.util.context.yelloSnackbar
 import com.example.ui.base.BindingActivity
