+++ app/src/main/java/com/el/yello/presentation/main/profile/info/ProfileFriendAdapter.kt
@@ -7,7 +7,6 @@ import androidx.recyclerview.widget.RecyclerView
 import com.el.yello.R
 import com.el.yello.databinding.ItemProfileFriendsListBinding
 import com.el.yello.databinding.ItemProfileUserInfoBinding
-import com.el.yello.presentation.main.profile.ProfileViewModel
 import com.example.domain.entity.ProfileUserModel
 import com.example.ui.view.ItemDiffCallback
