+++ app/src/main/java/com/el/yello/presentation/main/profile/info/ProfileFriendDeleteBottomSheet.kt
@@ -7,7 +7,6 @@ import androidx.lifecycle.flowWithLifecycle
 import androidx.lifecycle.lifecycleScope
 import com.el.yello.R
 import com.el.yello.databinding.FragmentProfileDeleteBottomSheetBinding
-import com.el.yello.presentation.main.profile.ProfileViewModel
 import com.el.yello.util.Utils.setImageOrBasicThumbnail
 import com.example.ui.base.BindingBottomSheetDialog
 import com.example.ui.fragment.toast
