+++ app/src/main/java/com/el/yello/presentation/main/profile/info/ProfileFriendItemBottomSheet.kt
@@ -6,7 +6,6 @@ import android.view.View
 import androidx.fragment.app.activityViewModels
 import com.el.yello.R
 import com.el.yello.databinding.FragmentProfileItemBottomSheetBinding
-import com.el.yello.presentation.main.profile.ProfileViewModel
 import com.el.yello.util.Utils.setImageOrBasicThumbnail
 import com.example.ui.base.BindingBottomSheetDialog
 import com.example.ui.view.setOnSingleClickListener
