+++ app/src/main/java/com/el/yello/presentation/main/profile/univmod/UnivProfileModViewModel.kt
@@ -1,13 +1,12 @@
-package com.el.yello.presentation.main.profile.mod
+package com.el.yello.presentation.main.profile.univmod
 
 import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
-import com.el.yello.presentation.main.profile.info.ProfileFragment.Companion.TYPE_HIGH_SCHOOL
-import com.el.yello.presentation.main.profile.info.ProfileFragment.Companion.TYPE_MIDDLE_SCHOOL
-import com.el.yello.presentation.main.profile.mod.UnivProfileModViewModel.Companion.TEXT_NONE
+import com.el.yello.presentation.main.profile.info.ProfileFragment.Companion.TYPE_UNIVERSITY
 import com.example.domain.entity.ProfileModRequestModel
-import com.example.domain.entity.onboarding.HighSchoolList
+import com.example.domain.entity.onboarding.GroupList
+import com.example.domain.entity.onboarding.SchoolList
 import com.example.domain.repository.OnboardingRepository
 import com.example.domain.repository.ProfileRepository
 import com.example.ui.view.UiState
@@ -21,7 +20,7 @@ import javax.inject.Inject
 import kotlin.math.ceil
 
 @HiltViewModel
-class SchoolProfileModViewModel @Inject constructor(
+class UnivProfileModViewModel @Inject constructor(
     private val profileRepository: ProfileRepository,
     private val onboardingRepository: OnboardingRepository
 ) : ViewModel() {
@@ -32,26 +31,26 @@ class SchoolProfileModViewModel @Inject constructor(
     private val _getIsModValidState = MutableStateFlow<UiState<Unit>>(UiState.Empty)
     val getIsModValidState: StateFlow<UiState<Unit>> = _getIsModValidState
 
-    private val _getSchoolListState = MutableStateFlow<UiState<HighSchoolList>>(UiState.Empty)
-    val getSchoolListState: StateFlow<UiState<HighSchoolList>> = _getSchoolListState
-
-    private val _getSchoolGroupIdResult = MutableSharedFlow<Boolean>()
-    val getSchoolGroupIdResult: SharedFlow<Boolean> = _getSchoolGroupIdResult
-
     private val _postToModProfileResult = MutableSharedFlow<Boolean>()
     val postToModProfileResult: SharedFlow<Boolean> = _postToModProfileResult
 
+    private val _getUnivListState = MutableStateFlow<UiState<SchoolList>>(UiState.Empty)
+    val getUnivListState: StateFlow<UiState<SchoolList>> = _getUnivListState
+
+    private val _getUnivGroupIdListState = MutableStateFlow<UiState<GroupList>>(UiState.Empty)
+    val getUnivGroupIdListState: StateFlow<UiState<GroupList>> = _getUnivGroupIdListState
+
     val lastModDate = MutableLiveData("")
     val school = MutableLiveData("")
-    val grade = MutableLiveData("")
-    val classroom = MutableLiveData("")
+    val subGroup = MutableLiveData("")
+    val admYear = MutableLiveData("")
 
     var isModAvailable = true
     var isChanged = false
 
-    private lateinit var myUserData: ProfileModRequestModel
+    lateinit var myUserData: ProfileModRequestModel
 
-    val classroomList = listOf("1", "2", "3", "4", "5", "6", "7", "8", "9", "10", "11", "12", "13", "14", "15", "16", "17", "18", "19", "20")
+    val studentIdList = listOf(24, 23, 22, 21, 20, 19, 18, 17, 16, 15)
 
     private var currentPage = -1
     private var isPagingFinish = false
@@ -64,7 +63,8 @@ class SchoolProfileModViewModel @Inject constructor(
     }
 
     fun resetStateVariables() {
-        _getSchoolListState.value = UiState.Empty
+        _getUnivListState.value = UiState.Empty
+        _getUnivGroupIdListState.value = UiState.Empty
     }
 
     fun getUserDataFromServer() {
@@ -75,14 +75,14 @@ class SchoolProfileModViewModel @Inject constructor(
                         _getUserDataResult.emit(false)
                         return@launch
                     }
-                    if (profile.groupType == TYPE_HIGH_SCHOOL || profile.groupType == TYPE_MIDDLE_SCHOOL) {
+                    if (profile.groupType == TYPE_UNIVERSITY) {
                         school.value = profile.groupName
-                        grade.value = profile.groupAdmissionYear.toString()
-                        classroom.value = profile.subGroupName
+                        subGroup.value = profile.subGroupName
+                        admYear.value = profile.groupAdmissionYear.toString()
                     } else {
                         school.value = TEXT_NONE
-                        grade.value = TEXT_NONE
-                        classroom.value = TEXT_NONE
+                        subGroup.value = TEXT_NONE
+                        admYear.value = "24"
                     }
                     myUserData = ProfileModRequestModel(
                         profile.name,
@@ -125,56 +125,13 @@ class SchoolProfileModViewModel @Inject constructor(
         }
     }
 
-    fun getSchoolListFromServer(searchText: String) {
-        if (isPagingFinish) return
-        viewModelScope.launch {
-            onboardingRepository.getHighSchoolList(
-                searchText,
-                ++currentPage
-            )
-                .onSuccess { schoolList ->
-                    if (schoolList == null) {
-                        _getSchoolListState.value = UiState.Empty
-                        return@launch
-                    }
-                    totalPage = ceil((schoolList.totalCount * 0.1)).toInt() - 1
-                    if (totalPage == currentPage) isPagingFinish = true
-                    _getSchoolListState.value = UiState.Success(schoolList)
-                }
-                .onFailure { t ->
-                    _getSchoolListState.value = UiState.Failure(t.message.toString())
-                }
-        }
-    }
-
-    fun getSchoolGroupIdFromServer() {
-        viewModelScope.launch {
-            onboardingRepository.getGroupHighSchool(
-                school.value ?: return@launch,
-                classroom.value ?: return@launch,
-            )
-                .onSuccess { data ->
-                    if (data == null) {
-                        _getSchoolGroupIdResult.emit(false)
-                        return@launch
-                    }
-                    myUserData.groupId = data.groupId
-                    _getSchoolGroupIdResult.emit(true)
-                    postNewProfileToServer()
-                }
-                .onFailure {
-                    _getSchoolGroupIdResult.emit(false)
-                }
-        }
-    }
-
-    private fun postNewProfileToServer() {
+    fun postNewProfileToServer() {
         viewModelScope.launch {
             if (!::myUserData.isInitialized) {
                 _postToModProfileResult.emit(false)
                 return@launch
             }
-            myUserData.groupAdmissionYear = grade.value?.toInt() ?: return@launch
+            myUserData.groupAdmissionYear = admYear.value?.toInt() ?: return@launch
 
             profileRepository.postToModUserData(myUserData)
                 .onSuccess {
@@ -184,5 +141,54 @@ class SchoolProfileModViewModel @Inject constructor(
                     _postToModProfileResult.emit(false)
                 }
         }
+    }
+
+    fun getUnivListFromServer(searchText: String) {
+        if (isPagingFinish) return
+        viewModelScope.launch {
+            onboardingRepository.getSchoolList(
+                searchText,
+                ++currentPage
+            )
+                .onSuccess { schoolList ->
+                    if (schoolList == null) {
+                        _getUnivListState.value = UiState.Empty
+                        return@launch
+                    }
+                    totalPage = ceil((schoolList.totalCount * 0.1)).toInt() - 1
+                    if (totalPage == currentPage) isPagingFinish = true
+                    _getUnivListState.value = UiState.Success(schoolList)
+                }
+                .onFailure { t ->
+                    _getUnivListState.value = UiState.Failure(t.message.toString())
+                }
+        }
+    }
+
+    fun getUnivGroupIdListFromServer(searchText: String) {
+        viewModelScope.launch {
+            _getUnivGroupIdListState.value = UiState.Loading
+            onboardingRepository.getGroupList(
+                ++currentPage,
+                school.value ?: return@launch,
+                searchText,
+            )
+                .onSuccess { groupList ->
+                    if (groupList == null || groupList.totalCount == 0) {
+                        _getUnivGroupIdListState.value = UiState.Empty
+                        return@launch
+                    }
+                    totalPage = ceil((groupList.totalCount * 0.1)).toInt() - 1
+                    if (totalPage == currentPage) isPagingFinish = true
+                    _getUnivGroupIdListState.value = UiState.Success(groupList)
+                }
+                .onFailure { t ->
+                    _getUnivGroupIdListState.value = UiState.Failure(t.message.toString())
+                }
+        }
+    }
+
+    companion object {
+        const val TEXT_NONE = "-"
     }
 }
