+++ app/src/main/res/layout/activity_profile_school_mod.xml
@@ -7,7 +7,7 @@
 
         <variable
             name="vm"
-            type="com.el.yello.presentation.main.profile.mod.SchoolProfileModViewModel" />
+            type="com.el.yello.presentation.main.profile.schoolmod.SchoolProfileModViewModel" />
 
         <import type="android.view.View" />
