+++ app/src/main/res/layout/activity_profile_univ_mod.xml
@@ -7,7 +7,7 @@
 
         <variable
             name="vm"
-            type="com.el.yello.presentation.main.profile.mod.UnivProfileModViewModel" />
+            type="com.el.yello.presentation.main.profile.univmod.UnivProfileModViewModel" />
 
         <import type="android.view.View" />
