+++ app/src/main/res/layout/fragment_profile_delete_bottom_sheet.xml
@@ -9,7 +9,7 @@
 
         <variable
             name="vm"
-            type="com.el.yello.presentation.main.profile.ProfileViewModel" />
+            type="com.el.yello.presentation.main.profile.info.ProfileViewModel" />
     </data>
 
     <androidx.constraintlayout.widget.ConstraintLayout
