+++ app/src/main/res/layout/item_profile_user_info.xml
@@ -10,7 +10,7 @@
 
         <variable
             name="vm"
-            type="com.el.yello.presentation.main.profile.ProfileViewModel" />
+            type="com.el.yello.presentation.main.profile.info.ProfileViewModel" />
     </data>
 
     <androidx.constraintlayout.widget.ConstraintLayout
